Update to EccsGuiBuilder, drop EccsWindowHelper
EccsWindowHelper got deleted and replaced by EccsGuiBuilder.
This commit fully switches to the new framework.

Additionally it will update the Readme.
Delete the mod CheeseMenu (as it is no longer needed).
Name keys inside of 'English_gui' are updated to use the full prefix. Less potential collision with other mods.