Update to 0.91.4 API
This commit does the bare minimum to support the latest update's API changes. This is primarily placing-rules and prefab generation code.

Instead of using one prefab class per component, now a generic prefab generator can generate them all. For that code-info values in the component definition files are being used.

The two components using ThroughPanelSegmentDisplay had the same custom data interface - it has been extracted, so that placing rules can generically apply to both of them. In case that this has to be undone a parent-interface with only the size could be used.

Added constants for default values shared by component code and generators.