Rework ram edit GUI
Reworks the layout usages, using the new EccsGuiBuilder layout helpers.
Now one can press ENTER in the text field to load the file.
Will activate the input field by default when opening it and keeps it activated on button/enter-key presses.
Sets a default Y-Position for the window. The center of the screen is now above the input-field.